+++ travel-api/transport-service/src/main/java/org/microarchitecturovisco/transport/controllers/reservations/ReservationRequest.java
@@ -6,6 +6,7 @@ import java.io.Serializable;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
+import java.math.BigDecimal;
 
 @Builder
 @Data
@@ -30,7 +31,7 @@ public class ReservationRequest implements Serializable {
     private List<UUID> departureLocationIdsByBus;
     private List<UUID> arrivalLocationIds;
 
-    private float price;
+    private BigDecimal price;
     private Integer hotelId;
     private List<Long> roomReservationsIds;
     private List<UUID> transportReservationsIds;
